Backend changes

+++ backend/src/main/java/com/motivamate/backend/controller/TaskRequest.java
@@ -1,24 +1,23 @@
 package com.motivamate.backend.controller;
 
 public class TaskRequest {
+
     private String title;
-    private int userId;
-    
-    
+    private Long userId;
+
     public String getTitle() {
         return title;
     }
 
-    // Getters and setters
     public void setTitle(String title) {
         this.title = title;
     }
-    public int getUserId() {
+
+    public Long getUserId() {
         return userId;
     }
-    public void setUserId(int userId) {
+
+    public void setUserId(Long userId) {
         this.userId = userId;
     }
-
-    
 }
